requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
pyTelegramBotAPI==4.16.1
python-telegram-bot==21.0.1
requests==2.31.0
sniffio==1.3.1
telebot==0.0.5
typing_extensions==4.10.0
urllib3==2.2.1
